Move services to extension method
Move the services for the API into an extension method.
